+++ bin/check-ssl-cert.rb
@@ -23,6 +23,7 @@
 # LICENSE:
 #   Jean-Francois Theroux <me@failshell.io>
 #   Nathan Williams <nath.e.will@gmail.com>
+#   Tom Braarup Cuykens <https://github.com/elfranne/>
 #   Released under the same terms as Sensu (the MIT license); see LICENSE
 #   for details.
 #
@@ -77,8 +78,18 @@ class CheckSSLCert < Sensu::Plugin::Check::CLI
          short: '-S',
          long: '--pass '
 
+  option :starttls,
+         description: 'Use StartTLS. OpenSSL 1.1.1 supports smtp, pop3, imap, ftp, xmpp, xmpp-server, irc, postgres, mysql, lmtp, nntp, sieve and ldap.',
+         long: '--starttls PROTOCOL '
+
   def ssl_cert_expiry
-    `openssl s_client -servername #{config[:servername]} -connect #{config[:host]}:#{config[:port]} < /dev/null 2>&1 | openssl x509 -enddate -noout`.split('=').last
+    `openssl s_client -servername #{config[:servername]} -connect #{config[:host]}:#{config[:port]} \
+    < /dev/null 2>&1 | openssl x509 -enddate -noout`.split('=').last
+  end
+
+  def ssl_cert_starttls
+    `openssl s_client -servername #{config[:servername]} -connect #{config[:host]}:#{config[:port]} -starttls #{config[:starttls]} \
+    < /dev/null 2>&1 | openssl x509 -enddate -noout`.split('=').last
   end
 
   def ssl_pem_expiry
@@ -102,15 +113,21 @@ class CheckSSLCert < Sensu::Plugin::Check::CLI
       end
     end
     config[:servername] = config[:host] unless config[:servername]
+
+    @starttls_supported_protocols = %w[smtp pop3 imap ftp xmpp xmpp-server irc postgres mysql lmtp nntp sieve ldap]
+    unless @starttls_supported_protocols.include?(config[:starttls]) # rubocop: disable Style/GuardClause
+      unknown "Unsupported StartTLS protocol #{config[:starttls]}" if config[:starttls]
+    end
   end
 
   def run
     validate_opts
-
     expiry = if config[:pem]
                ssl_pem_expiry
              elsif config[:pkcs12]
                ssl_pkcs12_expiry
+             elsif config[:starttls]
+               ssl_cert_starttls
              else
                ssl_cert_expiry
              end