+++ bin/check-ssl-crl.rb
@@ -65,7 +65,7 @@ class CheckSSLCRL < Sensu::Plugin::Check::CLI
   def run
     validate_opts
 
-    next_update = OpenSSL::X509::CRL.new(open(config[:url]).read).next_update
+    next_update = OpenSSL::X509::CRL.new(open(config[:url]).read).next_update # rubocop:disable Security/Open
     minutes_until = seconds_to_minutes(Time.parse(next_update.to_s) - Time.now)
 
     critical "#{config[:url]} - Expired #{minutes_until.abs} minutes ago" if minutes_until < 0