+++ bin/check-ssl-qualys.rb
@@ -1,6 +1,4 @@
 #!/usr/bin/env ruby
-# encoding: UTF-8
-
 #  check-ssl-qualys.rb
 #
 # DESCRIPTION:
@@ -138,7 +136,8 @@ class CheckSSLQualys < Sensu::Plugin::Check::CLI
                ssl_check(true)
              end
       return json if json['status'] == 'READY'
-      if json['endpoints'] && json['endpoints'].is_a?(Array)
+
+      if json['endpoints']&.is_a?(Array)
         p "endpoints: #{json['endpoints']}" if config[:debug]
         # The api response sometimes has low eta (which seems unrealistic) from
         # my tests that can be 0 or low numbers which would imply it is done...